msf-deployed/requirements.txt
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
transformers>=4.0.0
streamlit>=1.0.0
crewai>=0.1.0
crewai_tools>=0.1.0
PyGithub>=1.55
openai>=1.0.0
serper-dev>=0.1.0
